requirements.txt
black
flake8
flake8-docstrings
gmplot
isort
joblib
matplotlib
motmetrics
mypy
numpy
pandas
pillow
pycocotools
pylint
pytest
scikit-image
sphinx
sphinx-rtd-theme
tabulate
toml
torch
torchvision
tqdm

git+git://github.com/scalabel/scalabel.git